requirements.txt
fastapi
uvicorn[standard]
sqlmodel
websockets
python-dotenv